Compose --scale-luminance and --cap-luminance (scale, then clip)
LuminanceControl is now a struct of two optional stages instead of an
either/or enum: scale_max normalizes the whole image off its measured
peak, cap hard-clips what still exceeds it afterward. The common HDR
wallpaper failure — sane color levels under crazy white peaks — wants
both: a generous scale plus a tight cap.

Bonus from the rework: the declared luminances.max is now the honest
post-treatment ceiling in all cases — a scale that turns out to be a
no-op still declares the measured content peak instead of the source's
(often absurd) declaration, so compositor tone mapping gets real data
even from a plain --scale-luminance probe run.

Verified on the scRGB JXR: peak 1068.75 -> scale 800 -> cap 300,
declared max 300.

+++ src/app.rs
@@ -46,7 +46,7 @@ use crate::surfaces::{place, upload, upload_tiled, WireRgb8};
 
 /// Image identity for deduplication: path + luminance treatment (the same
 /// file capped differently for different outputs is different pixels).
-pub type ImageKey = (PathBuf, Option<(u8, u64)>);
+pub type ImageKey = (PathBuf, Option<(u64, u64)>);
 
 /// The dedup key for `spec`'s image, if it has one.
 pub fn image_key(spec: &OutputSpec) -> Option<ImageKey> {
